los .java seguros

+++ Practica3_Biblioteca/src/main/java/biblioteca/PortatilController.java
@@ -2,6 +2,8 @@ package biblioteca;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -23,14 +25,20 @@ public class PortatilController {
 		return portatiles_repo.findAll();
 	}
 	@RequestMapping(method=RequestMethod.POST)
-	public ResponseEntity<Portatil> addPortatil(@RequestBody Portatil portatil){
-		portatiles_repo.save(portatil);
-		return new ResponseEntity<>(portatil, HttpStatus.CREATED);
+	public ResponseEntity<Portatil> addPortatil(@RequestBody Portatil portatil,HttpSession sesion){
+		if(((sesion!=null)&&(sesion.getAttribute("admin") != null))&&((Boolean)sesion.getAttribute("admin"))){
+			portatiles_repo.save(portatil);
+			return new ResponseEntity<>(portatil, HttpStatus.CREATED);
+		}else{
+			return null;
+		}
 	}
 	
 	@RequestMapping(value = "/{id}", method= RequestMethod.DELETE)
-	public void deletePortatiles(@PathVariable long id){
-		portatiles_repo.delete(id);
+	public void deletePortatiles(@PathVariable long id, HttpSession sesion){
+		if(((sesion!=null)&&(sesion.getAttribute("admin") != null))&&((Boolean)sesion.getAttribute("admin"))){
+			portatiles_repo.delete(id);
+		}
 	}
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
@@ -39,8 +47,10 @@ public class PortatilController {
 	}
 	
 	@RequestMapping(value = "/{id}", method=RequestMethod.PUT)
-	public void modificar(@PathVariable long id, @RequestBody Portatil portatil){
-		portatiles_repo.setEstado(id,portatil.getEstado());
-		portatiles_repo.setCaracteristicas(id,portatil.getCaracteristicas());
+	public void modificar(@PathVariable long id, @RequestBody Portatil portatil,HttpSession sesion){
+		if(((sesion!=null)&&(sesion.getAttribute("admin") != null))&&((Boolean)sesion.getAttribute("admin"))){
+			portatiles_repo.setEstado(id,portatil.getEstado());
+			portatiles_repo.setCaracteristicas(id,portatil.getCaracteristicas());
+		}
 	}
 }
